sve prebaceno na ajaks deluje da sve radi

+++ 6. Realizacija/projekat/app/Controllers/Admin.php
@@ -46,18 +46,20 @@ class Admin extends Korisnik{
         $recenzijaModel=new RecenzijaModel();
         $time=new Time('now', 'Europe/Belgrade');
         $db->transStart();
-        $pisacModel->insert([
-            "IdKor"=>$idkor,
-            "PocetakKarijere"=>$time->toDateString()
-        ]);
-        $recenzijaModel->where('IdKor', $idkor)->delete();
-        $recenzentModel->delete($idkor);
-        $oblast=$oblastModel->find($idobl);
-        $oblastModel->update($idobl, [
-            "BrojRecenzenata"=>$oblast->BrojRecenzenata-1
-        ]);
+        $recenzent=$recenzentModel->find($idkor);
+        if($recenzent!=null){
+            $pisacModel->insert([
+                "IdKor"=>$idkor,
+                "PocetakKarijere"=>$time->toDateString()
+            ]);
+            $recenzijaModel->where('IdKor', $idkor)->delete();
+            $recenzentModel->delete($idkor);
+            $oblast=$oblastModel->find($idobl);
+            $oblastModel->update($idobl, [
+                "BrojRecenzenata"=>$oblast->BrojRecenzenata-1
+            ]);
+        }
         $db->transComplete();
-        return redirect()->back();
     }
     
     public function odobri($idobl, $idkor){
@@ -69,30 +71,71 @@ class Admin extends Korisnik{
         $oblast=$oblastModel->find($idobl);
         $brojRecenzenata=$oblast->BrojRecenzenata;
         $db->transStart();
-        if($brojRecenzenata<10){
-            $recenzentModel->insert([
-                "IdKor"=>$idkor,
-                "BrojZavrsenihRecenzija"=>0,
-                "IdObl"=>$idobl
-            ]);
-            $pisacModel->delete($idkor);
-            $oblast=$oblastModel->find($idobl);
-            $oblastModel->update($idobl, [
-                "BrojRecenzenata"=>$oblast->BrojRecenzenata+1
-            ]);
-            $zahtevModel->where("IdKor", $idkor)->delete();
-        }
-        else{
-            $this->session->setFlashdata("poruka", "Dostignut je maksimalan broj recenzenata za ovu oblast");
-            return redirect()->back();
+        $zahtev=$zahtevModel->where("IdObl", $idobl)->where("IdKor", $idkor)->first();
+        if($zahtev!=null){
+            if($brojRecenzenata<10){
+                $recenzentModel->insert([
+                    "IdKor"=>$idkor,
+                    "BrojZavrsenihRecenzija"=>0,
+                    "IdObl"=>$idobl
+                ]);
+                $pisacModel->delete($idkor);
+                $oblast=$oblastModel->find($idobl);
+                $oblastModel->update($idobl, [
+                    "BrojRecenzenata"=>$oblast->BrojRecenzenata+1
+                ]);
+                $zahtevModel->where("IdKor", $idkor)->delete();
+            }
+            else{
+                echo "Dostignut je maksimalan broj recenzenata za ovu oblast";
+            }
         }
         $db->transComplete();
-        return redirect()->back();
     }
     
     public function odbaci($idobl, $idkor) {
+        $db=db_connect();
         $zahtevModel=new ZahtevModel();
+        $db->transStart();
         $zahtevModel->where("IdObl", $idobl)->where("IdKor", $idkor)->delete();
-        return redirect()->back();
+        $db->transComplete();
+    }
+    
+    public function osveziZahteve($idobl){
+        $zahtevModel=new ZahtevModel();
+        $pisacModel=new PisacModel();
+        $recenzentModel=new RecenzentModel();
+        $ocenaModel=new OcenaModel();
+        $zahtevi=$zahtevModel->where("IdObl", $idobl)->findAll();
+        $pisci=$pisacModel->pisciZaZahteve($zahtevi);
+        $recenzenti=$recenzentModel->where("IdObl", $idobl)->findAll();
+        $prosecneOcene=$ocenaModel->prosecneOceneKorisnika($pisci, $recenzenti);
+        $i=0;
+        foreach($prosecneOcene as $prosecnaOcena){
+            $broj="";
+            $recenzent=false;
+            if($prosecnaOcena["brojRecenzija"]!=null){
+                $recenzent=true;
+                $broj=$broj.$prosecnaOcena["brojRecenzija"];
+            }
+            echo "<tr>
+                <td><a href='". site_url($this->getController()."/pregledTekstova/{$prosecnaOcena["id"]}")."'>{$prosecnaOcena["username"]}</a></td>
+                <td>{$prosecnaOcena["ocena"]}</td>
+                <td>{$broj}</td>";
+            echo "<script>parametri2[{$i}]=".$prosecnaOcena['id'].";</script>";
+            if($recenzent){
+                echo "<td>&nbsp;</td>
+                    <td>&nbsp;</td>
+                    <td><button class='btn btn-sm btn-warning vrati' data-index='{$i}'>Vrati</button></td>
+                </tr>";
+            }
+            else{
+                echo "<td><button class='btn btn-sm btn-success odobri' data-index='{$i}'>Odobri</button></td>
+                    <td><button class='btn btn-sm btn-danger odbaci' data-index='{$i}'>Odbaci</button></td>
+                    <td>&nbsp;</td>
+                </tr>";
+            }
+            $i++;
+        }
     }
 }
